fix(iran-map-picker): move province dataset into lib with alias imports
The relative ./iran-map-data import broke @workspace/ui's node16/nodenext typecheck (TS2835) with implicit-any fallout (TS7006), while a .js-suffixed variant broke Turbopack. Follow the persian-holidays-data pattern instead: data lives in lib/ on both sides, imported via package alias internally and @/lib in the consumer mirror - which also normalizes this file's stray @/ui/* prefixes to the standard @/hooks and @/lib aliases used by every other item.
