+++ apps/web/content/docs/components/iran-map-picker.mdx
@@ -22,7 +22,13 @@ Click or keyboard-navigate the map; the selected province shows below.
   sourceName="iran-map-picker"
   sourceTitle="components/ui/iran-map-picker.tsx"
   sourceStep="Copy the component into your project"
-  extraSources={[{ name: "iran-map-data", title: "components/ui/iran-map-data.ts", label: "Copy the province data file into your project" }]}
+  extraSources={[
+    {
+      name: "iran-map-data",
+      title: "lib/iran-map-data.ts",
+      label: "Copy the province data file into your project",
+    },
+  ]}
 />
 
 ## Usage
