+++ apps/web/registry.json
@@ -822,8 +822,8 @@
           "type": "registry:ui"
         },
         {
-          "path": "registry/base/ui/iran-map-data.ts",
-          "type": "registry:ui"
+          "path": "registry/base/lib/iran-map-data.ts",
+          "type": "registry:lib"
         }
       ]
     },
