+++ apps/web/registry/base/ui/iran-map-picker.tsx
@@ -2,10 +2,10 @@
 
 import * as React from "react"
 
-import { useControllableState } from "@/ui/hooks/use-controllable-state"
-import { cn } from "@/ui/lib/utils"
+import { useControllableState } from "@/hooks/use-controllable-state"
+import { cn } from "@/lib/utils"
 
-import { IRAN_MAP_VIEWBOX, IRAN_PROVINCE_SHAPES } from "./iran-map-data"
+import { IRAN_MAP_VIEWBOX, IRAN_PROVINCE_SHAPES } from "@/lib/iran-map-data"
 
 export interface IranMapPickerProps {
   /** The selected province ISO code, e.g. "IR-23". Use when controlled. */
